Nova estrutura para o projeto
