+++ src/br/edu/ifspcaraguatatuba/socket/Main.java
@@ -1,0 +1,17 @@
+package br.edu.ifspcaraguatatuba.socket;
+
+import javax.swing.JFrame;
+
+import br.edu.ifspcaraguatatuba.socket.view.StartChat;
+
+public class Main {
+
+	public static void main(String[] args) {
+		
+		
+		JFrame start = new StartChat();
+		start.setVisible(true);
+		
+	}
+
+}
